[refactor] Divide project in modules
divide the project in 4 modules:

- utils: useful utilities
- tagger: core feature of tagger
- ai: interaction with AI (chatgpt, mistralai ...)
- core: core modules (as custom_log.py)

Rename functions for clarity and move some of them.
