+++ .idea/misc.xml
@@ -4,4 +4,7 @@
     <option name="sdkName" value="Python 3.11 (Smart Tagger)" />
   </component>
   <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (Smart Tagger)" project-jdk-type="Python SDK" />
+  <component name="PythonCompatibilityInspectionAdvertiser">
+    <option name="version" value="3" />
+  </component>
 </project>
