+++ ai/__init__.py
@@ -1,0 +1,3 @@
+from .chatgpt import ChatGptAi, ChatGPTModel
+from .mistral import MistralAi, MistralAiModel
+from .abstract import Prompt, AbstractPromptProvider, AiEnum
